+++ packages/react/src/use-chat.ts
@@ -32,6 +32,7 @@ export type UseChatHelpers<UI_MESSAGE extends UIMessage> = {
   | 'stop'
   | 'resumeStream'
   | 'addToolResult'
+  | 'addToolOutput'
   | 'addToolApprovalResponse'
   | 'status'
   | 'messages'
@@ -128,7 +129,11 @@ export function useChat<UI_MESSAGE extends UIMessage = UIMessage>({
     error,
     resumeStream: chatRef.current.resumeStream,
     status,
-    addToolResult: chatRef.current.addToolResult,
+    /**
+     * @deprecated Use `addToolOutput` instead.
+     */
+    addToolResult: chatRef.current.addToolOutput,
+    addToolOutput: chatRef.current.addToolOutput,
     addToolApprovalResponse: chatRef.current.addToolApprovalResponse,
   };
 }